:truck: Move pyo3 functions under src/python/adapters.rs
Removing all pyo3 wrapper code from src/lib.rs, and putting them in src/python/adapters.rs instead to more clearly separate Rust and Python bindings. Also created a path_to_stream function to isolate the object_store code from the ndarray conversion code.
