+++ src/python/mod.rs
@@ -1,0 +1,2 @@
+/// Adapter interface from Rust to Python
+pub mod adapters;
